+++ tests/protocols/gmpv214/system/test_authentication.py
@@ -17,7 +17,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from ...gmpv214 import Gmpv214TestCase
-from ...gmpv208.entities.authentication import (
+from ...gmpv208.system.authentication import (
     GmpAuthenticateTestMixin,
     GmpDescribeAuthTestMixin,
     GmpModifyAuthTestMixin,
